Created database controller layer.
This change was made to allow the future creation of additional
database adapters. This change starts to make alterations that
will make it easier to connect different database types. The
future goal will be to have data move as specified in the
following outline:

Layer 1: ChatterBot
  - This layer interacts with users and apis. It handles a simple
api that allow responce data to be entered.

Layer 2: Storage Controller
  - This layer recieves raw text or json data from layer 1.
  - Here, raw data is converted into Statement objects.
  - Statement objects can be modified and passed to Layer 3.
  - Statement objects can also be retreived from Layer 3.

Layer 3: Database Adapter
  - The database adapter layer is an implementation of a
simple interface that specifies basic operations that are
required to use a type of database. These methods include
getting and writing data.